Fix Everything

+++ app/Models/Coordination.php
@@ -12,8 +12,6 @@ class Coordination extends Model
 
 
     protected $fillable = [
-        "sender_id",
-        "receiver_id",
         "title",
         "description",
         "start_time",
@@ -31,43 +29,36 @@ class Coordination extends Model
             $coordination->forwards()->delete();
 
             $coordination->followups->each(function ($followup) {
-                if ($followup->attachment) {
-                    Storage::delete($followup->attachment);
-                }
-
-                if ($followup->proof) {
-                    Storage::delete($followup->proof);
-                }
-
+                if ($followup->attachment) Storage::delete($followup->attachment);
+                if ($followup->proof) Storage::delete($followup->proof);
                 $followup->forwards()->delete();
-
+                if (method_exists($followup, 'scores')) {
+                    $followup->scores()->delete();
+                }
                 $followup->delete();
             });
         });
     }
 
-    public function sender()
+    public function senders()
     {
-        return $this->belongsTo(User::class, 'sender_id');
+        return $this->belongsToMany(User::class, 'coordination_users', 'coordination_id', 'sender_id')
+            ->withPivot('receiver_id')
+            ->withTimestamps();
     }
 
-    public function receiver()
+    public function receivers()
     {
-        return $this->belongsTo(User::class, 'receiver_id');
+
+        return $this->belongsToMany(User::class, 'coordination_users', 'coordination_id', 'receiver_id')
+            ->withPivot('sender_id')
+            ->withTimestamps();
     }
+
+
     public function followups()
     {
         return $this->hasMany(FollowupCoordination::class, 'coordination_id');
-    }
-
-    public function forwards()
-    {
-        return $this->hasMany(ForwardCoordination::class, 'coordination_id', 'id');
-    }
-
-    public function forwarder()
-    {
-        return $this->belongsTo(User::class, 'forwarded_by', 'id');
     }
 
     public function forwardedUsers()
@@ -77,6 +68,18 @@ class Coordination extends Model
             'forward_coordinations',
             'coordination_id',
             'forwarded_to'
-        );
+        )->withPivot('forwarded_by')
+            ->withTimestamps();
+    }
+
+    public function forwards()
+    {
+        return $this->hasMany(ForwardCoordination::class, 'coordination_id');
+    }
+
+
+    public function coordinationUsers()
+    {
+        return $this->hasMany(CoordinationUser::class, 'coordination_id');
     }
 }
